Make PDB building step optional.

+++ db_builder.py
@@ -124,16 +124,17 @@ def create_or_update_sqlite(args: argparse.Namespace):
         sqlite_conn.commit()
 
         # Set up the PDB<->uniprot DB
-        print("Doing PDB<->UniProt")
-        cursor.execute('DROP TABLE IF EXISTS pdb_tmp;')
-        cursor.execute('CREATE TABLE pdb_tmp (pdb_id text, uniprot_id text);')
-        cursor.executemany("INSERT INTO pdb_tmp(pdb_id, uniprot_id) values (?,?)",
-                           get_id_mappings(args.download_pdb, 'pdb'))
-        cursor.execute('DROP INDEX IF EXISTS pdb_index;')
-        cursor.execute('CREATE UNIQUE INDEX pdb_index ON pdb_tmp (pdb_id, uniprot_id);')
-        cursor.execute('DROP TABLE IF EXISTS pdb;')
-        cursor.execute('ALTER TABLE pdb_tmp RENAME TO pdb;')
-        sqlite_conn.commit()
+        if args.rebuild_pdb:
+            print("Doing PDB<->UniProt")
+            cursor.execute('DROP TABLE IF EXISTS pdb_tmp;')
+            cursor.execute('CREATE TABLE pdb_tmp (pdb_id text, uniprot_id text);')
+            cursor.executemany("INSERT INTO pdb_tmp(pdb_id, uniprot_id) values (?,?)",
+                               get_id_mappings(args.download_pdb, 'pdb'))
+            cursor.execute('DROP INDEX IF EXISTS pdb_index;')
+            cursor.execute('CREATE UNIQUE INDEX pdb_index ON pdb_tmp (pdb_id, uniprot_id);')
+            cursor.execute('DROP TABLE IF EXISTS pdb;')
+            cursor.execute('ALTER TABLE pdb_tmp RENAME TO pdb;')
+            sqlite_conn.commit()
 
 
 if __name__ == '__main__':
@@ -152,6 +153,11 @@ if __name__ == '__main__':
                         action='store_true',
                         dest='download_pdb',
                         help='Force re-download the PDB index before processing')
+    parser.add_argument('-n', '--no-pdb',
+                        action='store_false',
+                        default=True,
+                        dest='rebuild_pdb',
+                        help='Reload the PDB data. Default turned off.')
     args = parser.parse_args()
 
     create_or_update_sqlite(args)
